Refactor : 포인트 및 멤버쉽 호출 로직 추가

+++ src/main/java/com/example/hungrypangproject/domain/member/dto/request/SaveMemberRequest.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,5 +15,5 @@ public class SaveMemberRequest {
     private String address;
     private String phoneNo;
     private String password;
-    private Long totalPoint;
+    private BigDecimal totalPoint;
 }
