+++ src/main/java/com/example/hungrypangproject/domain/member/dto/respons/SearchMemberResponse.java
@@ -5,6 +5,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 public class SearchMemberResponse {
@@ -13,7 +15,7 @@ public class SearchMemberResponse {
     private final String nickName;
     private final String address;
     private final String phoneNo;
-    private final Long pointBalance;
+    private final BigDecimal pointBalance;
 
     public static SearchMemberResponse register(Member member) {
         return new SearchMemberResponse(
